% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_helpers.R
\name{charvec_to_formula}
\alias{charvec_to_formula}
\title{charvec_to_formula}
\usage{
charvec_to_formula(chr)
}
\arguments{
\item{chr}{character vector}
}
\value{
a formula
}
\description{
charvec_to_formula
}
\keyword{internal}
