% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_excel.R
\name{finish_excel_wb}
\alias{finish_excel_wb}
\title{finish excel workbook}
\usage{
finish_excel_wb(wb, wb_name)
}
\arguments{
\item{wb}{wb}

\item{wb_name}{wb name}
}
\value{
an excel file
}
\description{
finish excel workbook
}
