% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_helpers.R
\name{flex_create_headers}
\alias{flex_create_headers}
\title{flex_create_headers}
\usage{
flex_create_headers(df, word_vec)
}
\arguments{
\item{df}{a data frame}

\item{word_vec}{word vec}
}
\value{
data frame
}
\description{
create data frame of header specifications for flextable::set_header_df
}
\keyword{internal}
