% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_helpers.R
\name{flex_mergev1}
\alias{flex_mergev1}
\title{flex_mergev1}
\usage{
flex_mergev1(x, df)
}
\arguments{
\item{x}{a flextable}

\item{df}{data frame}
}
\value{
a flextable
}
\description{
flex_mergev1
}
\keyword{internal}
