% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_helpers.R
\name{flex_mergev1_formula}
\alias{flex_mergev1_formula}
\title{flex_mergev1_formula}
\usage{
flex_mergev1_formula(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a formula
}
\description{
flex_mergev1_formula
}
\keyword{internal}
