% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_helpers.R
\name{flex_set_headers}
\alias{flex_set_headers}
\title{charvec_to_formula}
\usage{
flex_set_headers(x, df, word_vec)
}
\arguments{
\item{x}{flextable}

\item{df}{data frame}

\item{word_vec}{word}
}
\value{
a flextable
}
\description{
charvec_to_formula
}
\keyword{internal}
