% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{format_currency}
\alias{format_currency}
\title{format currency}
\usage{
format_currency(tbl, ..., symbol = "¥", digits = 0)
}
\arguments{
\item{tbl}{dataframe}

\item{...}{tidyselect.}

\item{digits}{integer. trailing digits}
}
\value{
dataframe
}
\description{
formats a monetary value with the currency symbol
}
