% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{format_percent}
\alias{format_percent}
\title{format percentage}
\usage{
format_percent(tbl, ..., digits = 0)
}
\arguments{
\item{tbl}{dataframe}

\item{...}{tidyselect}

\item{digits}{integer. trailing digits}
}
\value{
dataframe
}
\description{
formats a number as a percentage. default selection uses a helper to detect numeric columns
in between -1 and 1.
}
