% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{format_red}
\alias{format_red}
\title{format_red}
\usage{
format_red(flextbl, col, clr = "firebrick3")
}
\arguments{
\item{flextbl}{flextable}

\item{col}{string. col name}

\item{clr}{color}
}
\value{
flextable
}
\description{
format_red
}
\keyword{internal}
