% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_piped_name.R
\name{get_piped_name}
\alias{get_piped_name}
\title{get piped name}
\usage{
get_piped_name(object, default_name = "Table")
}
\arguments{
\item{object}{an object}

\item{default_name}{string}
}
\value{
string
}
\description{
this function captures the name of an object piped into a function, and returns as a string. Powers the automatic naming found in presenter.
}
\details{
The behavior doesn't work as expected when examples or vignettes are being knitted, so instead
you can specify a default output.
}
\examples{

### on local machine the output will be "iris" for each example

### works if the object is piped or given as an argument
iris \%>\%
get_piped_name()

get_piped_name(iris)

### can even extract name from multistep pipes
iris \%>\%
dplyr::select(1:3) \%>\%
get_piped_name()

### can be placed inside other functions to capture the name and save it

find_name <- function(x){
 get_piped_name() -> new_name

 new_name
}

iris \%>\%
dplyr:select(1:3) \%>\%
find_name()
}
