% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{is_percentage}
\alias{is_percentage}
\title{is_percentage}
\usage{
is_percentage(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical
}
\description{
is_percentage
}
\examples{

library(presenter)

c(.1, 0, .5) \%>\%
is_percentage
}
\keyword{internal}
