% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_excel.R
\name{make_excel}
\alias{make_excel}
\title{Create Excel}
\usage{
make_excel(
  df,
  last_id_col = NULL,
  header_word = NULL,
  widths = 13,
  random_color_seed = 1
)
}
\arguments{
\item{df}{data frame}

\item{last_id_col}{index of last id col}

\item{header_word}{character vector of header words}

\item{widths}{col widths}

\item{random_color_seed}{seed for random color scheme}
}
\value{
an excel file
}
\description{
Create Excel
}
