% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_excel.R
\name{make_excel_wb}
\alias{make_excel_wb}
\title{create excel wb}
\usage{
make_excel_wb(
  wb = NULL,
  object,
  last_id_col = NULL,
  header_word = NULL,
  widths = 13,
  random_color_seed = 1
)
}
\arguments{
\item{wb}{wb}

\item{object}{object}

\item{last_id_col}{index of last id col}

\item{header_word}{character vector of header words}

\item{widths}{col widths}

\item{random_color_seed}{seed for random color scheme}
}
\value{
excel wb object
}
\description{
create excel wb
}
