% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flextable.R
\name{make_flextable}
\alias{make_flextable}
\title{Make Flextable}
\usage{
make_flextable(df, header_words = NULL, last_id_col = NULL)
}
\arguments{
\item{df}{data frame}

\item{header_words}{header words}

\item{last_id_col}{last id col}
}
\value{
a flextable
}
\description{
Make Flextable
}
