% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pivot_table.R
\name{make_pivot_table}
\alias{make_pivot_table}
\title{Make pivot table}
\usage{
make_pivot_table(
  tbl,
  col1,
  col2,
  show_percentages = T,
  tbl_nm = NULL,
  tabyl_denominator = "all",
  theme = c("zebra_blue", "zebra_gold", "tron", "vader", "vanilla", "booktabs",
    "alafoli")
)
}
\arguments{
\item{tbl}{a data frame to pivot}

\item{col1}{unqouted col 1}

\item{col2}{unquoted col 2}

\item{show_percentages}{logical; show percentages with numbers}

\item{tbl_nm}{string to name table. If not given, automatically defaults to table name.}

\item{tabyl_denominator}{denominator to use when calculating percentages}

\item{theme}{string to choose a predefined theme}
}
\value{
a flextable
}
\description{
Make pivot table
}
