% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_powerpoint.R
\name{make_powerpoint}
\alias{make_powerpoint}
\title{Send Table to Powerpoint}
\usage{
make_powerpoint(
  tbl,
  output_file = NULL,
  layout = "Two Content",
  master = "Office Theme",
  show = TRUE
)
}
\arguments{
\item{tbl}{a data frame,  flextable, or a list thereof}

\item{output_file}{path to existing ppt}

\item{layout}{master layout}

\item{master}{master theme}

\item{show}{logical to open the ppt}
}
\value{
none
}
\description{
A table can be piped in to this function to be sent to a new ppt slide. Can be called with no arguments,
then a new ppt is created named after the table. If output_file is specified, table is sent to a new slide
on an existing powerpoint.
}
