% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simple_excel.R
\name{make_simple_excel}
\alias{make_simple_excel}
\title{Make a simple excel}
\usage{
make_simple_excel(object, output_file = NULL)
}
\arguments{
\item{object}{a data frame}

\item{output_file}{string to name excel file. If not given, defaults to name of table.}
}
\value{
an .xlsx file
}
\description{
Export a file to excel with minimal formatting and minimal effort.
}
