% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_summary.R
\name{pivot_summary}
\alias{pivot_summary}
\title{Pivot Summary}
\usage{
pivot_summary(sumr, ...)
}
\arguments{
\item{sumr}{A tibble}

\item{...}{an optional tidyselect specification of grouping columns to pivot}
}
\value{
a tibble
}
\description{
Wrapper around base r `t()` that returns a tibble.
Transposes a data frame, intended for use on the output of a  dplyr::summarize operation
}
\details{
for an ungrouped summarize, ... argument can be left empty
for a grouped summarize, use column names or tidyselect to pivot the group names. Multiple groups
will be concatenated before pivoting.
}
