% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_excel.R,
%   R/style_header_sequence_openxlsx2.R
\name{style_header_sequence}
\alias{style_header_sequence}
\title{style_header_sequence}
\usage{
style_header_sequence(header_word, object, id_col_range, wb, sheet)

style_header_sequence(header_word, object, id_col_range, wb, sheet)
}
\arguments{
\item{header_word}{A string to match against column names}

\item{object}{A data.frame-like object}

\item{id_col_range}{Columns to exclude from styling}

\item{wb}{A Workbook object from openxlsx2}

\item{sheet}{A string specifying the worksheet name}
}
\value{
excel headers

Applies styles to matching headers in an openxlsx2 workbook
}
\description{
style_header_sequence

style_header_sequence
}
\keyword{internal}
