% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_piped_name.R
\name{get_piped_name}
\alias{get_piped_name}
\title{get piped name}
\usage{
get_piped_name(objName, default_name = "Title")
}
\arguments{
\item{objName}{an object}

\item{default_name}{default name if object name is not able to be accessed in case of a long pipe}
}
\value{
string
}
\description{
this function captures the name of a data frame piped into a function as a string. Powers the automatic naming found in presenter.
}
\details{
citation: <https://michaelbarrowman.co.uk/post/getting-a-variable-name-in-a-pipeline/>
}
\keyword{internal}
