% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flextable.R
\name{make_flextable}
\alias{make_flextable}
\title{Make Flextable}
\usage{
make_flextable(
  df,
  header_words = NULL,
  last_id_col = NULL,
  merge_col_indices = NULL,
  dbl_digits = 2,
  theme = c("zebra_blue", "zebra_gold", "tron", "vader", "vanilla", "booktabs",
    "alafoli")
)
}
\arguments{
\item{df}{data frame}

\item{header_words}{header words. Takes a character vector of header words. will be automatically generate via a heuristic if left NULL. can be completely disabled by the string "disable"}

\item{last_id_col}{last id col}

\item{merge_col_indices}{merge specific column indices}

\item{dbl_digits}{integer. how many trailing digits should be displayed on dbls}

\item{theme}{string to choose a preselected theme}
}
\value{
a flextable
}
\description{
Turns a data frame into a flextable
}
