% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simple_excel.R
\name{make_simple_excel}
\alias{make_simple_excel}
\title{Make a simple excel}
\usage{
make_simple_excel(object, show = TRUE)
}
\arguments{
\item{object}{a data frame or list thereof}

\item{show}{logical. open excel upon completion?}
}
\value{
an .xlsx file
}
\description{
Export a file to excel with minimal formatting and minimal effort.
}
