% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_excel.R
\name{style_header_sequence}
\alias{style_header_sequence}
\title{style_header_sequence}
\usage{
style_header_sequence(header_word, object, id_col_range, wb, sheet)
}
\arguments{
\item{header_word}{header words}

\item{object}{object}

\item{id_col_range}{id col range}

\item{wb}{wb}

\item{sheet}{sheet}
}
\value{
excel headers
}
\description{
style_header_sequence
}
\keyword{internal}
